Initial DHFSS building side implementation
